requirements.txt
APScheduler==3.6.3
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
canvasapi==2.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
pycodestyle==2.7.0
python-telegram-bot==13.7
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
tzlocal==3.0
urllib3==1.26.7
